+++ CHANGELOG.md
@@ -19,3 +19,4 @@ Release summary...
 - A user can no longer enter an invalid C++ variable name.
 - XRC files now include a comment warning that regeneration will eliminate any hand edits.
 - `Generate XRC files...` on Tools menu now launches a dialog allowing you to choose between combined or individually generated XRC files.
+- wxStatusBar `fields` property now supports unique width and style for each field.